+++ example/lib/src/pubspec.dart
@@ -1,6 +1,11 @@
-// Do not modify.
+// Generated file. Do not modify.
 //
 // This file is generated using the build_pubspec package.
 // For more information, go to: https://pub.dev/packages/build_pubspec
-
-const String packageVersion = '0.1.2';
+const List<String> authors = [
+  'Vince Varga <vince.varga@serial.dev>',
+];
+const String description =
+    '''This is an example description. Great job! Awesome package.''';
+const String name = 'build_pubspec_example';
+const String version = '0.1.2';